+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.0
+ - [ADDED] Support for Bootstrap3. (#1)
+ - [CHANGED] Optional field can now be hidden. (#2)
+ - [CHANGED] Card title is now configurable. (#5)
+ - [CHANGED] Headers in Trello are now configurable. (#4)
+ - [CHANGED] Reload URL is now configurable. (#3)
+ - [FIXED] Show error, when no lists are available. (#6)
+
 ## 1.1.0
  - Categories are now sorted alphabetically
  - Added validation for E-Mails and unfilled fields